Kasperngbergan/pat 25 implement fft computation for audio buffer

+++ NUCLEOH723ZG/src/apps/ics52000/app_main.c
@@ -30,9 +30,9 @@ void app_main(void)
                         peak = a;
                 }
                 float rms  = sqrtf((float)sum_sq / (fmt.samples_per_block));
-				float dbfs = 20.0f * log10f(rms / 2147483648.0f + 1e-20f);
+								float dbfs = 20.0f * log10f(rms / 2147483648.0f + 1e-20f);
                 // printf("%8ld %8ld %6d\t", (long)rms, (long)peak, (int)dbfs);
-				printf("%ld", (long)dbfs);
+				printf("%.1f", dbfs);
 				if (i + 1 < fmt.mic_count) printf(",");
             }
             printf("\r\n");
